# cmd/vulnerability-db-api-cli/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Vulnerability Database API
  description: API for the Vulnerability Database
  version: "1.0"
servers:
- url: https://vulnerabilitydb.vulcan.com
paths:
...
  /findings:
    get:
      tags:
      - findings
      summary: List findings findings
      description: |-
        This method allows to list findings by using different filters.
        When querying for open findings in a specific time range, it returns every finding that has been found in that time, regardless if that finding has also been fixed in that span of time.
        When querying for fixed findings in a specific time range, it returns every finding that has been fixed in that time and it's fixed at the end of that time range, so a finding that is fixed and then reopened in that range of time will not be returned.
      operationId: findings#List findings
      parameters:
      - name: tag
        in: query
        description: Targets tag filter
        allowEmptyValue: true
        schema:
          type: string
          description: Targets tag filter
          example: Molestiae at voluptates accusamus.
        example: Optio repellat enim eos voluptatum esse et.
      - name: status
        in: query
        description: Finding status filter. Can be 'OPEN' or 'FIXED'.
        allowEmptyValue: true
        schema:
          type: string
          description: Finding status filter. Can be 'OPEN' or 'FIXED'.
          example: Quis quae quae est ea et ad.
        example: Quisquam reiciendis quia dolor corporis.
      - name: minScore
        in: query
        description: Minimum issues score filter
        allowEmptyValue: true
        schema:
          type: integer
          description: Minimum issues score filter
          example: 977222862732478355
          format: int64
        example: 595567487579622434
      - name: maxScore
        in: query
        description: Maximum issues score filter
        allowEmptyValue: true
        schema:
          type: integer
          description: Maximum issues score filter
          example: 6929356274958660579
          format: int64
        example: 796405614657449144
      - name: atDate
        in: query
        description: Allows to get a list of findings for a specific date (YYY-MM-DD).
          This filter has preference over minDate and maxDate.
        allowEmptyValue: true
        schema:
          type: string
          description: Allows to get a list of findings for a specific date (YYY-MM-DD).
            This filter has preference over minDate and maxDate.
          example: Placeat facere voluptates voluptatem veritatis.
        example: Soluta cum molestias doloribus ea.
      - name: minDate
        in: query
        description: Minimum date filter (YYY-MM-DD). This filter only applies when
          used in conjunction with status filter.
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter (YYY-MM-DD). This filter only applies when
            used in conjunction with status filter.
          example: Repudiandae officia adipisci.
        example: Velit officia sint.
      - name: maxDate
        in: query
        description: Maximum date filter (YYY-MM-DD). This filter only applies when
          used in conjunction with status filter.
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter (YYY-MM-DD). This filter only applies when
            used in conjunction with status filter.
          example: Nam rem fugit.
        example: Dolorum velit unde sint eveniet nobis.
      - name: sortBy
        in: query
        description: 'Sorting criteria. Supported fields: score, -score (for descending
          order)'
        allowEmptyValue: true
        schema:
          type: string
          description: 'Sorting criteria. Supported fields: score, -score (for descending
            order)'
          example: Eaque quaerat ab.
        example: Voluptatem voluptatem quam corrupti quam.
      - name: page
        in: query
        description: Number of page requested
        allowEmptyValue: true
        schema:
          type: integer
          description: Number of page requested
          example: 3954292642761844345
          format: int64
        example: 5133827017924758047
      - name: size
        in: query
        description: Number of results per page
        allowEmptyValue: true
        schema:
          type: integer
          description: Number of results per page
          example: 2812677772936176575
          format: int64
        example: 4743216853249427978
      - name: identifier
        in: query
        description: Findings identifier filter
        allowEmptyValue: true
        schema:
          type: string
          description: Findings identifier filter
          example: Incidunt qui recusandae omnis quia.
        example: Quos corporis in.
      - name: identifiermatch
        in: query
        description: Exact match in identifier filter
        allowEmptyValue: true
        schema:
          type: boolean
          description: Exact match in identifier filter
          example: false
        example: false
      - name: identifiers
        in: query
        description: A list of asset identifiers
        allowEmptyValue: true
        schema:
          type: string
          description: A list of asset identifiers
          example: Labore eum.
        example: Explicabo tempore dolorum nisi facilis dicta.
      - name: labels
        in: query
        description: List of issue labels to filter by
        allowEmptyValue: true
        schema:
          type: string
          description: List of issue labels to filter by
          example: Corporis vitae ut.
        example: Est quis eaque aut velit.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingCollection'
              example:
              - details: TLSv1 is enabled and the server supports at least one cipher
                id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                impact_details: The remote service encrypts traffic using an older
                  version of TLS.
                issue:
                  cwe_id: 216
                  description: |-
                    The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                    number of cryptographic design flaws. Modern implementations of TLS 1.0
                    mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                    designed against these flaws and should be used whenever possible.

                    PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                    2018, except for POS POI terminals (and the SSL/TLS termination
                    points to which they connect) that can be verified as not being
                    susceptible to any known exploits.
                  id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                  labels:
                  - Web,TLS
                  recommendations:
                  - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                    1.0.
                  reference_links:
                  - http://www.example.com
                  summary: TLS Version 1.0 Protocol Detection
                resources:
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                score: 7
                source:
                  component: vulcan-tls
                  id: 8502fa58-6066-4890-8039-6cd936885763
                  instance: 6087cba3-fade-4fee-81ca-7459e87de098
                  name: vulcan
                  options: '{"timeout":60}'
                  time: "2019-06-08 11:16:40"
                status: OPEN
                target:
                  id: 8d14bf4b-2cb1-43f2-9c23-f7985c63858a
                  identifier: www.example.com
                  tags:
                  - sdrn:adevinta:team:security
              - details: TLSv1 is enabled and the server supports at least one cipher
                id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                impact_details: The remote service encrypts traffic using an older
                  version of TLS.
                issue:
                  cwe_id: 216
                  description: |-
                    The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                    number of cryptographic design flaws. Modern implementations of TLS 1.0
                    mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                    designed against these flaws and should be used whenever possible.

                    PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                    2018, except for POS POI terminals (and the SSL/TLS termination
                    points to which they connect) that can be verified as not being
                    susceptible to any known exploits.
                  id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                  labels:
                  - Web,TLS
                  recommendations:
                  - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                    1.0.
                  reference_links:
                  - http://www.example.com
                  summary: TLS Version 1.0 Protocol Detection
                resources:
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                score: 7
                source:
                  component: vulcan-tls
                  id: 8502fa58-6066-4890-8039-6cd936885763
                  instance: 6087cba3-fade-4fee-81ca-7459e87de098
                  name: vulcan
                  options: '{"timeout":60}'
                  time: "2019-06-08 11:16:40"
                status: OPEN
                target:
                  id: 8d14bf4b-2cb1-43f2-9c23-f7985c63858a
                  identifier: www.example.com
                  tags:
                  - sdrn:adevinta:team:security
              - details: TLSv1 is enabled and the server supports at least one cipher
                id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                impact_details: The remote service encrypts traffic using an older
                  version of TLS.
                issue:
                  cwe_id: 216
                  description: |-
                    The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                    number of cryptographic design flaws. Modern implementations of TLS 1.0
                    mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                    designed against these flaws and should be used whenever possible.

                    PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                    2018, except for POS POI terminals (and the SSL/TLS termination
                    points to which they connect) that can be verified as not being
                    susceptible to any known exploits.
                  id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                  labels:
                  - Web,TLS
                  recommendations:
                  - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                    1.0.
                  reference_links:
                  - http://www.example.com
                  summary: TLS Version 1.0 Protocol Detection
                resources:
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                score: 7
                source:
                  component: vulcan-tls
                  id: 8502fa58-6066-4890-8039-6cd936885763
                  instance: 6087cba3-fade-4fee-81ca-7459e87de098
                  name: vulcan
                  options: '{"timeout":60}'
                  time: "2019-06-08 11:16:40"
                status: OPEN
                target:
                  id: 8d14bf4b-2cb1-43f2-9c23-f7985c63858a
                  identifier: www.example.com
                  tags:
                  - sdrn:adevinta:team:security
...
  /sources:
    get:
      tags:
      - sources
      summary: List sources sources
      operationId: sources#List sources
      parameters:
      - name: page
        in: query
        description: Number of page requested
        allowEmptyValue: true
        schema:
          type: integer
          description: Number of page requested
          example: 4094373485487189010
          format: int64
        example: 5380516567400390768
      - name: size
        in: query
        description: Number of results per page
        allowEmptyValue: true
        schema:
          type: integer
          description: Number of results per page
          example: 8222056356488407934
          format: int64
        example: 6979530151777249718
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceCollection'
              example:
              - component: vulcan-tls
                id: 8502fa58-6066-4890-8039-6cd936885763
                instance: 6087cba3-fade-4fee-81ca-7459e87de098
                name: vulcan
                options: '{"timeout":60}'
                target_id: Consequuntur veniam.
                time: "2019-06-08 11:16:40"
              - component: vulcan-tls
                id: 8502fa58-6066-4890-8039-6cd936885763
                instance: 6087cba3-fade-4fee-81ca-7459e87de098
                name: vulcan
                options: '{"timeout":60}'
                target_id: Consequuntur veniam.
                time: "2019-06-08 11:16:40"
  /sources/{id}:
    get:
      tags:
      - sources
      summary: Get source sources
      operationId: sources#Get source
      parameters:
      - name: id
        in: path
        description: Source ID
        required: true
        schema:
          type: string
          description: Source ID
          example: Non sit quia sint.
        example: Sint repellendus.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
              example:
                component: vulcan-tls
                id: 8502fa58-6066-4890-8039-6cd936885763
                instance: 6087cba3-fade-4fee-81ca-7459e87de098
                name: vulcan
                options: '{"timeout":60}'
                target_id: Voluptas earum laboriosam.
                time: "2019-06-08 11:16:40"
...
  /sources/{id}/mttr:
    get:
      tags:
      - sources
      summary: Get source MTTR sources
      operationId: sources#Get source MTTR
      parameters:
      - name: id
        in: path
        description: Source ID
        required: true
        schema:
          type: string
          description: Source ID
          example: Tenetur dolores.
        example: Delectus itaque quasi explicabo.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                type: number
                example: 0.31628352
                format: float
              example: 0.6014624
...
components:
  schemas:
...
    Source:
      type: object
      properties:
        component:
          type: string
          description: Source component
          example: vulcan-tls
        id:
          type: string
          description: Source ID
          example: 8502fa58-6066-4890-8039-6cd936885763
        instance:
          type: string
          description: Source instance
          example: 6087cba3-fade-4fee-81ca-7459e87de098
        name:
          type: string
          description: Source name
          example: vulcan
        options:
          type: string
          description: Source options
          example: '{"timeout":60}'
        target_id:
          type: string
          description: Target ID
          example: Dicta quos.
        time:
          type: string
          description: Time on which the source scanned the target
          example: "2019-06-08 11:16:40"
      example:
        component: vulcan-tls
        id: 8502fa58-6066-4890-8039-6cd936885763
        instance: 6087cba3-fade-4fee-81ca-7459e87de098
        name: vulcan
        options: '{"timeout":60}'
        target_id: Voluptas minus illum enim.
        time: "2019-06-08 11:16:40"
    SourceCollection:
      type: array
      items:
        $ref: '#/components/schemas/Source'
      example:
      - component: vulcan-tls
        id: 8502fa58-6066-4890-8039-6cd936885763
        instance: 6087cba3-fade-4fee-81ca-7459e87de098
        name: vulcan
        options: '{"timeout":60}'
        target_id: Consequuntur veniam.
        time: "2019-06-08 11:16:40"
      - component: vulcan-tls
        id: 8502fa58-6066-4890-8039-6cd936885763
        instance: 6087cba3-fade-4fee-81ca-7459e87de098
        name: vulcan
        options: '{"timeout":60}'
        target_id: Consequuntur veniam.
        time: "2019-06-08 11:16:40"
      - component: vulcan-tls
        id: 8502fa58-6066-4890-8039-6cd936885763
        instance: 6087cba3-fade-4fee-81ca-7459e87de098
        name: vulcan
        options: '{"timeout":60}'
        target_id: Consequuntur veniam.
        time: "2019-06-08 11:16:40"
...
  securitySchemes:
    authorization_header_Authorization:
      type: apiKey
      description: Value must be 'TAG tag=mytag'
      name: Authorization
      in: header
...